fix bugs get coords

+++ Graph_algorithms/Graph_algorithms/Build_dots/Build_dots.cpp
@@ -19,7 +19,7 @@ namespace build_dots
 	Point getDot(Point mainDot, double angle, int i, double radius)
 	{
 		Point resPoint;
-		double helpY = radius*cos(pi*i*angle / 360.0);
+		double helpY = radius*cos(pi*double(i)*angle / 360.0);
 		resPoint.y = mainDot.y - helpY;
 		resPoint.x = sqrt(radius * radius - helpY * helpY);
 		return resPoint;
